==> 10-data-cleaning-validation.ipynb <==
import pandas as pd

df = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train.csv')

df.head(20)

df.shape

**In the training dataset, there are about 1.5 million rows and 11 columns.**

df.isna().sum(axis=0)

print('Percentage of missing values in each columns:')
df.isna().sum(axis=0)/df.shape[0]*100

# Loop through all columns to get unique values
for column in df.columns:
    num_unique_values = df[column].nunique()
    print(f"{column}: (Unique Count: {num_unique_values})")
    print()

# Find the rows without any missing values
df_nonan = df.dropna()
df_nonan.shape
