==> 30-feature-engineering.ipynb <==
# OUTPUTS of the feature engineering notebook
**products_train_UK_representations.csv**
- This file contains 3 columns: `id`, `price`: numeric, and `representation`: a numeric vector of dimension 768, representing all the information of the products.

**sessions_UK_finalized_len2.csv**
- This file contains 52199 rows × 2 columns: `next_item`, `prev_items`
- Only keep the rows where the number of previous items purchased >= 2, if it's longer than 2, only keep the latest 2 items

**sessions_UK_finalized_len3.csv**
- This file contains 25615 rows × 2 columns: `next_item`, `prev_items`
- Only keep the rows where the number of previous items purchased >= 3, if it's longer than 3, only keep the latest 3 items

df_UK_representations

sessions_train_UK_filtered_len2

sessions_train_UK_filtered_len3

import pandas as pd
from transformers import BertTokenizer, BertModel
import torch

df = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train.csv')

# find the rows where id is not unique
df[df.duplicated(subset=['id'], keep=False)].sort_values(by='id')

# Keep only locale = 'UK' for simplicity

df_UK = df[df['locale'] == 'UK']
df_UK[df_UK.duplicated(subset=['id'], keep=False)].sort_values(by='id')

df_UK.shape

df_UK['id'].nunique()

# Save the cleaned data
# df_UK.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK.csv', index=False)

df_UK

df_UK = df_UK.fillna('')

df_UK['description'] = (
    'title: ' + df_UK['title'] +
    '\nbrand: ' + df_UK['brand'] +
    '\ncolor: ' + df_UK['color'] +
    '\nsize: ' + df_UK['size'] +
    '\nmodel: ' + df_UK['model'] +
    '\nmaterial: ' + df_UK['material'] +
    '\nauthor: ' + df_UK['author'] +
    '\ndesc: ' + df_UK['desc']
)

df_UK = df_UK[['id', 'price', 'description']]
df_UK

print(df_UK['description'][0])

## dataset: sessions_train, for simplicity, only keep locale = 'UK' and items appears more than 20 times

sessions_train = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/sessions_train.csv')

sessions_train

sessions_train_UK = sessions_train[sessions_train['locale'] == 'UK']

sessions_train_UK.shape

from collections import Counter
import re

# First, let's concatenate prev_items and next_item into a single list
all_items = []
for index, row in sessions_train_UK.iterrows():
    # Use regular expression to match all alphanumeric sequences
    prev_items_list = re.findall(r"'([^']*)'", row['prev_items'])

    # Combine with the next_item
    combined_items = prev_items_list + [row['next_item']]
    all_items.extend(combined_items)

# Count the frequency of each item ID
item_counter = Counter(all_items)


# Extract items that appear several times
# HERE I SET THE THRESHOLD TO 20
items_with_multiple_occurrences = [item for item, count in item_counter.items() if count > 100]

len(items_with_multiple_occurrences)

# Now, filter the DataFrame to only keep rows where 'next_item' is in 'items_with_multiple_occurrences'
sessions_train_UK_filtered = sessions_train_UK[sessions_train_UK['next_item'].isin(items_with_multiple_occurrences)]
sessions_train_UK_filtered.shape

sessions_train_UK_filtered.drop('locale', axis=1, inplace=True)


sessions_train_UK_filtered

sessions_train_UK_filtered.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/sessions_train_UK_filtered.csv', index=False)

# Initialize an empty list to store the filtered prev_items
filtered_prev_items = []

# Loop over each row in the DataFrame
for index, row in sessions_train_UK_filtered.iterrows():
    # Use regular expression to match all alphanumeric sequences
    prev_items_list = re.findall(r"'([^']*)'", row['prev_items'])

    # Remove items not in items_with_multiple_occurrences
    filtered_list = [item for item in prev_items_list if item in items_with_multiple_occurrences]
    
    # Append the filtered list to the accumulator list
    filtered_prev_items.append(filtered_list)

# Assign the filtered lists to a new column in the DataFrame
sessions_train_UK_filtered['pre_items_filter'] = filtered_prev_items


sessions_train_UK_filtered

sessions_train_UK_filtered.drop('prev_items', axis=1, inplace=True)
sessions_train_UK_filtered.rename(columns={'pre_items_filter': 'prev_items'}, inplace=True)

sessions_train_UK_filtered

sessions_train_UK_filtered.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/sessions_train_UK_filtered.csv', index=False)

# Check the mean length of the filtered prev_items
lengths= sessions_train_UK_filtered['prev_items'].str.len()
stats = lengths.describe()

# Display the statistics
print(stats)

lengths.value_counts()

### Only keep the rows where the number of previous items purchased >= 2, if it's longer than 2, only keep the latest 2 items

sessions_train_UK_filtered_len2 = sessions_train_UK_filtered.copy()
sessions_train_UK_filtered_len2

# Define a function to keep only the latest 2 items if the length is greater than 2
def keep_latest_two(items):
    if len(items) >= 2:
        return items[-2:]
    else:
        return items

# Apply the function to the 'prev_items' column
sessions_train_UK_filtered_len2['prev_items'] = sessions_train_UK_filtered_len2['prev_items'].apply(keep_latest_two)

sessions_train_UK_filtered_len2

# Filter the DataFrame to only keep rows where the length of 'prev_items' is at least 2
sessions_train_UK_filtered_len2 = sessions_train_UK_filtered_len2[sessions_train_UK_filtered_len2['prev_items'].str.len() >= 2]
sessions_train_UK_filtered_len2.shape

sessions_train_UK_filtered_len2

sessions_train_UK_filtered_len2.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/sessions_UK_finalized_len2.csv', index=False)

### Only keep the rows where the number of previous items purchased >= 3, if it's longer than 3, only keep the latest 3 items

sessions_train_UK_filtered_len3 = sessions_train_UK_filtered.copy()
sessions_train_UK_filtered_len3

# Define a function to keep only the latest 2 items if the length is greater than 2
def keep_latest_three(items):
    if len(items) >= 3:
        return items[-3:]
    else:
        return items

# Apply the function to the 'prev_items' column
sessions_train_UK_filtered_len3['prev_items'] = sessions_train_UK_filtered_len3['prev_items'].apply(keep_latest_three)

sessions_train_UK_filtered_len3

# Filter the DataFrame to only keep rows where the length of 'prev_items' is at least 3
sessions_train_UK_filtered_len3 = sessions_train_UK_filtered_len3[sessions_train_UK_filtered_len3['prev_items'].str.len() >= 3]
sessions_train_UK_filtered_len3.shape

sessions_train_UK_filtered_len3

sessions_train_UK_filtered_len3.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/sessions_UK_finalized_len3.csv', index=False)

### Filter out the items information that are not in the items_with_multiple_occurrences

df_UK = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK_finalized.csv')

df_UK

# filter out the products which are not in the list items_with_multiple_occurrences

df_UK_filtered = df_UK[df_UK['id'].isin(items_with_multiple_occurrences)]
df_UK_filtered.shape

df_UK_filtered.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK_finalized.csv', index=False)

df_UK_filtered

df_UK_finalized = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK_finalized.csv')

df_UK_finalized

# Set the device to CPU
device = torch.device('cpu')

# Load the tokenizer and model
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
model = BertModel.from_pretrained('bert-base-uncased').to(device)
model.eval()  # Put the model in evaluation mode

# Initialize the list to hold representations
representations = []

# Determine the number of chunks to process
total = len(df_UK_finalized)
chunk_size = 100  # Set this to something that fits in your memory
num_chunks = total // chunk_size + (1 if total % chunk_size != 0 else 0)

# Process the dataframe in chunks
for i in range(num_chunks):
    start_index = i * chunk_size
    end_index = start_index + chunk_size
    
    # Perform tokenization for the current chunk
    inputs = tokenizer(df_UK_finalized['description'].iloc[start_index:end_index].tolist(),
                       padding=True,
                       truncation=True,
                       max_length=512,
                       return_tensors="pt").to(device)
    
    # Forward pass to get embeddings
    with torch.no_grad():
        outputs = model(**inputs)
    
    # Extract the pooled output
    paragraph_representations = outputs.pooler_output.detach().cpu().numpy().tolist()
    representations.extend(paragraph_representations)
    
    # Display the progress
    print(f"Chunk {i+1}/{num_chunks} processed. ({(i+1)/num_chunks:.2%} complete)")

# After processing all chunks, add the representations to the dataframe
df_UK_finalized['representation'] = representations


df_UK_finalized

df_UK_representations = df_UK_finalized[['id', 'price', 'representation']]
df_UK_representations.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK_representations.csv', index=False)

df_UK_representations
