==> 32-features-continue2.ipynb <==
import pandas as pd
import numpy as np

items = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/products_train_UK_representations.csv')

items.dtypes

items.head()

representation_list = items['representation'].apply(eval).tolist()

# Convert the list of lists to an array where each row is the representation
representations_array = np.array(representation_list)

representations_array

representations_array.shape

# Map items to indices
item_to_index = {item_id: index for index, item_id in enumerate(items['id'])}


df_index = pd.read_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/train_and_test_index.csv')

## Adjacency Matrix

# Initialize an adjacency matrix with zeros
adjacency_matrix = np.zeros((len(items), len(items)))

# Iterate over the training dataset to populate the adjacency matrix
for _, row in df_index.iterrows():

    adjacency_matrix[row['last_item_index'], row['next_item_index']] += 1  # Increment the weight for the link
    adjacency_matrix[row['last_second_index'], row['last_item_index']] += 1  # Increment the weight for the link

# transpose the matrix because I want to aggregate the features from the previous items
adjacency_matrix = adjacency_matrix.T

# Sum the elements of each row (axis=1 for rows)
row_sums = adjacency_matrix.sum(axis=1)

# Avoid division by zero by replacing zero row sums with 1
row_sums[row_sums == 0] = 1

# Divide each row by its sum to normalize it so that each row sums to 1
adjacency_matrix_normalized = adjacency_matrix / row_sums[:, np.newaxis]

### Aggregating the Features
Now we have the adjacency matrix, we can aggregate features. Since we're interested in the likelihood of purchasing the next_item based on prev_items, we want to collect features from source nodes (prev_items) to destination nodes (next_item). Thus, we should use A*X:

**Note: here we use the transpose of A, which is A.T, because we want to aggregate features from source nodes to destination nodes.**

# Now the row-normalized adjacency_matrix can be used to perform the aggregation
aggregated_representation = adjacency_matrix_normalized.dot(representations_array)

aggregated_representation.shape

combined_representation = np.hstack((representations_array, aggregated_representation))
combined_representation.shape

combined_price = items[['price', 'aggregated_price']].values
combined_price.shape

items['price'].describe()

combined_features = np.hstack((combined_representation, combined_price))

combined_features.shape

# Save the array to a file 
np.save('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/combined_features.npy', combined_features)

items

items_with_combined_features = items[['index', 'id']].copy()

items_with_combined_features['combined_features']=combined_features.tolist()

items_with_combined_features

items_with_combined_features.to_csv('/Users/stoneman/Library/CloudStorage/OneDrive-Vanderbilt/Social_Network_Analysis/Amazon_Recommandation_Project/data/items_with_combined_features.csv', index=False)
